Split action manager logic into normal and hijack (win32 only) action managers
