+++ tf2_bot_detector/PlatformSpecific/Shitty.h
@@ -2,5 +2,6 @@
 
 namespace tf2_bot_detector
 {
+	bool IsTF2Running();
 	void RequireTF2NotRunning();
 }
